refactor: extract fair status routes to routes/fairStatus.js
